+++ sky-server/src/main/java/com/sky/mapper/EmployeeMapper.java
@@ -56,5 +56,5 @@ public interface EmployeeMapper {
      * @return
      */
     @Select("select  * from  employee where  id = #{id}")
-    Employee getById(Integer id);
+    Employee getById(Long id);
 }
